refactor: consolidate i18n service
