+++ src/app/layout/language-switcher/language-switcher.spec.ts
@@ -1,7 +1,9 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { LanguageSwitcherComponent } from './language-switcher';
-import { I18nService } from '../../i18n/i18n.service';
+import { I18nService } from '../../core/i18n/i18n.service';
+import { I18nModule } from '../../core/i18n/i18n.module';
 
 describe('LanguageSwitcherComponent', () => {
   let component: LanguageSwitcherComponent;
@@ -10,7 +12,12 @@ describe('LanguageSwitcherComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [LanguageSwitcherComponent, RouterTestingModule]
+      imports: [
+        LanguageSwitcherComponent,
+        RouterTestingModule,
+        I18nModule,
+        HttpClientTestingModule
+      ]
     }).compileComponents();
 
     i18nService = TestBed.inject(I18nService);
@@ -20,6 +27,6 @@ describe('LanguageSwitcherComponent', () => {
 
   it('should update active language in I18nService when switching', () => {
     component.switchLanguage('en');
-    expect(i18nService.currentLang).toBe('en');
+    expect(i18nService.lang()).toBe('en');
   });
 });
